try to fix downstream build -- run ci
